+++ src/refactorings/extract/extract-type/extract-type.ts
@@ -78,6 +78,10 @@ function createVisitor(
     TSTypeAnnotation(path) {
       if (!selection.isInsidePath(path)) return;
 
+      // Since we visit nodes from parent to children, first check
+      // if a child would match the selection closer.
+      if (hasChildWhichMatchesSelection(path, selection)) return;
+
       const { typeAnnotation } = path.node;
 
       if (
@@ -114,4 +118,22 @@ function createVisitor(
       );
     }
   };
+}
+
+function hasChildWhichMatchesSelection(
+  path: t.NodePath,
+  selection: Selection
+): boolean {
+  let result = false;
+
+  path.traverse({
+    TSTypeAnnotation(childPath) {
+      if (!selection.isInsidePath(childPath)) return;
+
+      result = true;
+      childPath.stop();
+    }
+  });
+
+  return result;
 }